10/30

+++ src/app/components/member/petgiving-record-page/petgiving-record-page.component.ts
@@ -2,9 +2,8 @@ import { Component } from '@angular/core';
 import { FormBuilder, Validators } from '@angular/forms';
 import { MatDialog } from '@angular/material/dialog';
 import { AdoptService } from 'src/app/service/adopt.service';
-import { PetFormModel } from '../../interfaces/pet.interface';
-import { DomSanitizer, SafeUrl } from '@angular/platform-browser';
 import { DialogComponent } from '../../dialog/dialog.component';
+import { PetFormModel } from '../../interfaces/pet.interface';
 
 @Component({
   selector: 'app-petgiving-record-page',
@@ -20,27 +19,27 @@ export class PetgivingRecordPageComponent {
   selectedPetId = 0;
   petPhoto: string | undefined;
   modifyInfoList: PetFormModel[] = [];
+  selectedFile!: string | null;
 
   constructor(private formBuilder: FormBuilder, private service: AdoptService, public dialog: MatDialog) { }
 
   form = this.formBuilder.group({
-    name: ['', Validators.required],
+    name: ['', [Validators.required, Validators.pattern(/[\S]/)]],
     species: ['', Validators.required],
     city: ['', Validators.required],
-    type: ['', Validators.required],
-    size: ['', Validators.required],
-    color: ['', Validators.required],
-    age: ['', Validators.required],
+    type: ['', [Validators.required, Validators.pattern(/[\S]/)]],
+    size: ['', [Validators.required, Validators.pattern(/[\S]/)]],
+    color: ['', [Validators.required, Validators.pattern(/[\S]/)]],
+    age: ['', [Validators.required, Validators.pattern(/[\S]/)]],
     sex: ['', Validators.required],
     ligation: [false],
-    introduction: ['', Validators.required],
+    introduction: [''],
     photo: [''],
     affidavit: [false],
     followUp: [false],
     ageLimit: [false],
     parentalPermission: [false]
   })
-
 
   ngOnInit() {
     this.service.showAllCity().subscribe(response => {
@@ -52,53 +51,67 @@ export class PetgivingRecordPageComponent {
     });
     //show petgiving record
     //看要怎樣傳入id，可能登入之後做查詢存在service之後方便使用
-    this.selectedPetId = 0;
     this.service.showPetGivingRecord(this.userId).subscribe(response => {
       console.log(response.response), this.petList = response.response
     });
   }
 
   deletePetInfo() {
-    //重新查詢
-    this.service.deletePetInfo(this.selectedPetId).subscribe(response => {
-      if (response.statusCode === '0000') {
-        this.ngOnInit();
-        this.dialog.open(DialogComponent, {
-          data: { dialogMode: 'deleteSuccessDialog' }
+    const dialogRef = this.dialog.open(DialogComponent, {
+      data: { dialogMode: 'deleteDialog', petName: this.form.value.name }
+    });
+    dialogRef.afterClosed().subscribe(petName => {
+      //有確定刪除
+      if (petName !== undefined) {
+        console.log(petName);
+        this.service.deletePetInfo(this.selectedPetId).subscribe(response => {
+          if (response.statusCode === '0000') {
+            this.dialog.open(DialogComponent, {
+              data: { dialogMode: 'deleteSuccessDialog' }
+            });
+            //重新查詢
+            this.refresh();
+          } else {
+            this.dialog.open(DialogComponent, {
+              data: { dialogMode: 'deleteFailedDialog' }
+            });
+          }
         });
-      } else {
-        this.dialog.open(DialogComponent, {
-          data: { dialogMode: 'deleteFailedDialog' }
-        });
+
       }
     });
+
   }
 
   modifyPetInfo() {
+    console.log(this.form.value);
+    console.log(this.form);
     if (this.form.invalid) {
       this.dialog.open(DialogComponent, {
         data: { dialogMode: 'invalidInputDialog' }
       });
     } else if (this.form.valid) {
+      const formData = this.form.value;
+      console.log(formData);
       const petData: PetFormModel = {
         id: this.selectedPetId,
-        name: this.form.value.name || null,
-        species: this.form.value.species || null,
-        city: this.form.value.city || null,
-        type: this.form.value.type || null,
-        size: this.form.value.size || null,
-        color: this.form.value.color || null,
-        age: this.form.value.age || null,
-        sex: this.form.value.sex || null,
-        ligation: this.form.value.ligation === true ? 'Y' : 'N',
-        introduction: this.form.value.introduction!,
-        photo: null,//待解決
+        name: formData.name || null,
+        species: formData.species || null,
+        city: formData.city || null,
+        type: formData.type || null,
+        size: formData.size || null,
+        color: formData.color || null,
+        age: formData.age || null,
+        sex: formData.sex || null,
+        ligation: formData.ligation === true ? 'Y' : 'N',
+        introduction: formData.introduction!,
+        photo: this.selectedFile,
         postStatus: null,//沒用到
         publishDate: null,//沒用到
-        conditionAffidavit: this.form.value.affidavit === true ? 'Y' : 'N',
-        conditionFollowUp: this.form.value.followUp === true ? 'Y' : 'N',
-        conditionAgeLimit: this.form.value.ageLimit === true ? 'Y' : 'N',
-        conditionParentalPermission: this.form.value.parentalPermission === true ? 'Y' : 'N',
+        conditionAffidavit: formData.affidavit === true ? 'Y' : 'N',
+        conditionFollowUp: formData.followUp === true ? 'Y' : 'N',
+        conditionAgeLimit: formData.ageLimit === true ? 'Y' : 'N',
+        conditionParentalPermission: formData.parentalPermission === true ? 'Y' : 'N',
         //先預設userId = 2
         userId: this.userId
       };
@@ -108,10 +121,10 @@ export class PetgivingRecordPageComponent {
       this.service.updatePetInfo(this.modifyInfoList).subscribe(response => {
         //更新成功dialog
         if (response.statusCode === '0000') {
-          this.ngOnInit();
           this.dialog.open(DialogComponent, {
             data: { dialogMode: 'updateSuccessDialog' }
           });
+          this.refresh();
         } else {
           this.dialog.open(DialogComponent, {
             data: { dialogMode: 'updateFailedDialog' }
@@ -121,14 +134,23 @@ export class PetgivingRecordPageComponent {
     }
   }
 
+  refresh() {
+    this.service.showPetGivingRecord(this.userId).subscribe(response => {
+      console.log(response.response), this.petList = response.response
+    });
+    this.modifyInfoList = [];
+    this.selectedPetId = 0;
+  }
+
   selectPet(pet: PetFormModel[]) {
+    console.log(this.form);
     this.selectedPetId = pet[0].id!;
     console.log(pet[0]);
+    this.selectedFile = pet[0].photo;
     //將值放到畫面上
     this.form.patchValue({
       name: pet[0].name,
       //要再想想，因為pet[0].species是speciesName，不是speciesId，已在後端更改
-      //圖片問題
       species: pet[0].species,
       city: pet[0].city,
       type: pet[0].type,
@@ -147,4 +169,23 @@ export class PetgivingRecordPageComponent {
 
   }
 
+  async onPhotoChange(event: any) {
+    if (event.target.files.length > 0) {
+      const file = event.target.files[0];
+      this.selectedFile = await this.getBase64(file).then();
+      console.log(this.selectedFile);
+      console.log(this.selectedFile!.split(",")[1]);
+      this.selectedFile = this.selectedFile!.split(",")[1];//將前綴(blob:)拿掉
+      console.log(this.selectedFile);
+    }
+  }
+
+  getBase64(file: Blob) {
+    return new Promise((resolve, reject) => {
+      const reader = new FileReader();
+      reader.readAsDataURL(file);
+      reader.onload = () => resolve(reader.result);
+      reader.onerror = error => reject(error);
+    });
+  }
 }
